Fixed sync for modules in subdirectories more than 1 level deep.
The root cause was that CompilerOutputModuleCustomizer expected
ExtIdeaCompilerOutput#getMainClassesDir() to be not null.
In the case of modules that are inside other folders, the Gradle
IdeaProject creates a new module per folder. For example, for
the subproject libraries:library1, Gradle will return the
module libraries 'libraries' and 'library1'. The module 'libraries'
will not have any "main" classes folder, throwning an exception
when trying to set up the project.

Fixing this bug requires the following:
1. Check for null "main" classes folder in
   CompilerOutputModuleCustomizer
2. The most important issue here is that the exception was lost,
   the API being used to customize modules
   (ExternalSystemApiUtil#executeProjectChangeAction) throws away
   the exception. This CL replaces that API with WriteCommandAction,
   which returns any caught exception. We also log the exception
   with ERROR level, to let users know that something went wrong.

Fixes
https://code.google.com/p/android/issues/detail?id=71098

Change-Id: I43daca53c187aa192eb6db790b6c8c019708b048
(cherry picked from commit 36fe9237ba7d7f12754765a12205794fcd92f542)