Fixed bug where "SDK Manager" action in "Welcome" page is always disabled.
This happens with a new installation of Android Studio. When opening
Studio for the first time, Studio tries to find an Android SDK to use. If
is able to find one, it creates the appropriate Android SDKs (IntelliJ
SDK) but the "SDK Manager" is never enabled after that.

Change-Id: I07f058fa8f072c4e76b8a28f1044aab00f178f11
(cherry picked from commit 9868b1e64bacab96f7f009344d750a4d5c013179)