Fixes the automatic fix for 'android-L' SDK path bug.
CL https://android-review.googlesource.com/100326
only checked the path of the IAndroidTarget location, but that
is not enough. The location may return a correct value, but some
files in the SDK can still point to the wrong location,
android.jar being the most important one.

Another bug is that we were setting the SDK location outside of
a "write action" which is illegal (this is the result of
improper testing.)

Change-Id: I068c25e9621b58055354493c583b2a176a5b8fc3
(cherry picked from commit d04f1e033f8f50c38f46882fa70a17ddc91a3518)