Adds support for reproxing methods that throw exceptions
If the model is out of that but the IDE handles it by catching the
exception that calling a method that doesn't exist in the actual model,
reproxied objects need to behave in the same way.

This change also improves nested proxy objects testing.

Change-Id: I8ba8d1b74d532ff53eb1eca6f69ba56f3a77addd
(cherry picked from commit e25f85a25d35dcd595098325bbbbf4b3db27b0ed)