Fixed regression where projects wrapper was not created for projects using it.
This CL fixes a regression where the Gradle wrapper was not being
automatically created for projects that are configured to use the wrapper
but, for some reason, did not have the wrapper. This CL fixes that
by automatically creating the wrapper if necessary.

This CL also fixes another bug: Gradle sync warns the user that the
project is using the wrong version of Gradle only if the project
is using plug-in 0.13.0. In this CL we extended the check to all the
supported versions of the plugin: 0.12.+ and 0.13.0.

Change-Id: I0c2e4b2d86bc5ae4c6f0d87afd9daa4d70943d24
(cherry picked from commit 17a3b8f20fbc02fb5d7ebcd50d7ffefe8db9539b)