Now we don't offer to fix Gradle version for plug-in 0.12.+.
In CL https://android-review.googlesource.com/108094, we offer
to fix Gradle version (to v1.12) if the plug-in is version 0.12.+.
The problem is that 0.12.+ supports many versions of Gradle, and
if the project uses a valid version like 1.11, the IDE will still
offer to update the Gradle version, which is confusing.

Instead of implementing a complete map of Gradle versions vs.
plug-in versions, we just let sync fail if the project is using
the wrong version of Gradle. We already offer a quick-fix in
case sync fails due to a wrong Gradle version.

Now we only offer to fix Gradle version (to 2.1) if the project
is using plug-in 0.13.0.

Change-Id: I5b3727e271018263e4e36dbf0561b841cf5af082
(cherry picked from commit e3adc95e9c32e42b68dd9e81e5941e35e189d240)