Hyperlinks support at the build output
https://code.google.com/p/android/issues/detail?id=75901

The general idea was to combine two existing functionalities:
* show single line of text on multiple visual lines if available UI width
  is not enough (backed by MultilineTreeCellRenderer)
* correctly represent hyperlinks with an ability to configure custom
  click handlers at the IDE UI

IntelliJ code analysis showed that the later facility is supported by
SimpleColoredComponentClass and ColoredTreeCellRenderer. They are used
e.g. in 'Git Log' and 'File History' panels. So, what we did initially was enhanced
SimpleColoredComponent which knew how to display text on multiple lines if
necessary. Corresponding patch was offered to JetBrains -
http://youtrack.jetbrains.com/issue/IDEA-130051. However, they can apply
it not earlier than in two weeks because of their release cycle, that's why we
copy-pasted target classes into android plugin and enhanced them with the required
'wrap mode' facility. We're planning to switch to the platform code later when
JetBrains apply the patch.

Change-Id: Ie656235353493b71c085d8bc670b29224631924c